+++ doc/release-notes.md
@@ -3,3 +3,5 @@ Bitcoin ABC version 0.20.12 is now available from:
   <https://download.bitcoinabc.org/0.20.12/>
 
 This release includes the following features and fixes:
+ - The `unloadwallet` RPC is now synchronous, meaning that it blocks until the
+   wallet is fully unloaded.